assignment_07/pyproject.toml
[project]
name = "assignment-07"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
